+++ src/mcpdotnet/Client/IMcpClient.cs
@@ -107,6 +107,8 @@ public interface IMcpClient : IAsyncDisposable
 
     /// <summary>
     /// Sends a generic JSON-RPC request to the server.
+    /// NB! This is a temporary method that is available to send not yet implemented feature messages. 
+    /// Once all MCP features are implemented this will be made private, as it is purely a convenience for those who wish to implement features ahead of the library.
     /// </summary>
     /// <typeparam name="T">The expected response type.</typeparam>
     /// <param name="request">The JSON-RPC request to send.</param>
