+++ src/mcpdotnet/Protocol/Messages/JsonRpcError.cs
@@ -5,7 +5,7 @@ using System.Text.Json.Serialization;
 /// <summary>
 /// An error response message in the JSON-RPC protocol.
 /// </summary>
-public record JsonRpcError : IJsonRpcMessageWithId
+internal record JsonRpcError : IJsonRpcMessageWithId
 {
     /// <summary>
     /// JSON-RPC protocol version. Always "2.0".
