+++ src/mcpdotnet/Protocol/Transport/StdioTransport.cs
@@ -2,7 +2,6 @@
 namespace McpDotNet.Protocol.Transport;
 
 using System.Diagnostics;
-using System.Text;
 using System.Text.Json;
 using System.Threading;
 using McpDotNet.Protocol.Messages;
@@ -13,14 +12,29 @@ using McpDotNet.Utils.Json;
 /// </summary>
 public record StdioTransportOptions
 {
+    /// <summary>
+    /// The command to execute to start the server process.
+    /// </summary>
     public required string Command { get; set; }
 
+    /// <summary>
+    /// Arguments to pass to the server process.
+    /// </summary>
     public string[]? Arguments { get; set; } = Array.Empty<string>();
 
+    /// <summary>
+    /// The working directory for the server process.
+    /// </summary>
     public string? WorkingDirectory { get; set; }
 
+    /// <summary>
+    /// Environment variables to set for the server process.
+    /// </summary>
     public Dictionary<string, string>? EnvironmentVariables { get; set; }
 
+    /// <summary>
+    /// The timeout to wait for the server to shut down gracefully.
+    /// </summary>
     public TimeSpan ShutdownTimeout { get; init; } = TimeSpan.FromSeconds(5);
 }
