--- src/mcpdotnet/Protocol/Types/IResourceContents.cs
@@ -1,10 +1,0 @@
-﻿namespace McpDotNet.Protocol.Types;
-
-/// <summary>
-/// Interface for resource contents.
-/// </summary>
-public interface IResourceContents
-{
-    string Uri { get; }
-    string? MimeType { get; }
-}
