+++ src/mcpdotnet/Protocol/Types/PromptMessage.cs
