+++ src/mcpdotnet/Utils/Json/JsonSerializerOptionsExtensions.cs
@@ -5,7 +5,7 @@ namespace McpDotNet.Utils.Json;
 /// <summary>
 /// Extensions for configuring System.Text.Json serialization options for MCP.
 /// </summary>
-public static class JsonSerializerOptionsExtensions
+internal static class JsonSerializerOptionsExtensions
 {
     /// <summary>
     /// Configures JsonSerializerOptions with MCP-specific settings and converters.
@@ -16,8 +16,6 @@ public static class JsonSerializerOptionsExtensions
     {
         // Add custom converters
         options.Converters.Add(new JsonRpcMessageConverter());
-        options.Converters.Add(new ContentJsonConverter());
-        options.Converters.Add(new ResourceContentsJsonConverter());
         options.Converters.Add(new JsonStringEnumConverter());
 
         // Configure general options
